finished markdown migration
