+++ src/common/components/MarkDownParser/index.jsx
@@ -1,221 +1,20 @@
-/* eslint-disable no-loop-func */
-/* eslint-disable no-await-in-loop */
+/* eslint-disable react/prop-types */
+import React, { useEffect, useState } from 'react';
+import ReactMarkdown from 'react-markdown';
+import remarkGfm from 'remark-gfm';
+import remarkGemoji from 'remark-gemoji';
 import PropTypes from 'prop-types';
+import rehypeRaw from 'rehype-raw';
+import { Img } from '@chakra-ui/react';
+
 import useTranslation from 'next-translate/useTranslation';
-import { compiler } from 'markdown-to-jsx';
-import { Box, Checkbox, Link } from '@chakra-ui/react';
-import { PrismLight as SyntaxHighlighter } from 'react-syntax-highlighter';
-import jsx from 'react-syntax-highlighter/dist/cjs/languages/prism/jsx';
-import js from 'react-syntax-highlighter/dist/cjs/languages/prism/javascript';
-import css from 'react-syntax-highlighter/dist/cjs/languages/prism/css';
-import python from 'react-syntax-highlighter/dist/cjs/languages/prism/python';
-import bash from 'react-syntax-highlighter/dist/cjs/languages/prism/bash';
-
-import emoji from 'emoji-dictionary';
-import { useEffect, useState } from 'react';
+import {
+  BeforeAfter, Code, MDHeading, MDHr, MDLink, MDText, OnlyForBanner,
+} from './MDComponents';
 import { usePersistent } from '../../hooks/usePersistent';
-import tomorrow from './syntaxHighlighter/tomorrow';
-import BeforeAfterSlider from '../BeforeAfterSlider';
-import CallToAction from '../CallToAction';
 import Toc from './toc';
-import Heading from '../Heading';
-// import Anchor from './Anchor';
-// import ChakraHeading from '../Heading';
-import Text from '../Text';
 import ContentHeading from './ContentHeading';
-import OnlyFor from '../OnlyFor';
-
-// okaidia-tomorrow
-SyntaxHighlighter.registerLanguage('jsx', jsx);
-SyntaxHighlighter.registerLanguage('js', js);
-SyntaxHighlighter.registerLanguage('html', jsx);
-SyntaxHighlighter.registerLanguage('css', css);
-SyntaxHighlighter.registerLanguage('python', python);
-SyntaxHighlighter.registerLanguage('bash', bash);
-
-const Code = ({
-  inline, className, children, ...props
-}) => {
-  const match = /lang-(\w+)/.exec(className || '');
-
-  return !inline && match ? (
-    <SyntaxHighlighter
-      showLineNumbers
-      style={tomorrow}
-      language={match[1]}
-      PreTag="div"
-      {...props}
-    >
-      {String(children).replace(/\n$/, '')}
-    </SyntaxHighlighter>
-  ) : (
-    <code className={`${className} highlight`} {...props}>
-      {children}
-    </code>
-  );
-};
-
-function doWithDelay(timeout, doCallback) {
-  return new Promise((res) => {
-    setTimeout(() => {
-      doCallback();
-      res();
-    }, timeout);
-  });
-}
-
-const BeforeAfter = ({ before, after }) => {
-  const [delimerPersentPosition, setDelimerPercentPosition] = useState(50);
-  const animationDemo = () => {
-    setTimeout(async () => {
-      const PARTS = 50;
-      const timeSeconds = 0.1;
-      const borderMin = 35;
-      const delta = (delimerPersentPosition - borderMin) / PARTS;
-      const timeout = (timeSeconds / PARTS) * 1000;
-      let currentPosition = delimerPersentPosition;
-
-      for (let i = 1; i <= PARTS; i += 1) {
-        await doWithDelay(timeout, () => {
-          currentPosition -= delta;
-          setDelimerPercentPosition(currentPosition);
-        });
-      }
-      await doWithDelay(1000, () => {});
-      for (let i = 1; i <= PARTS; i += 1) {
-        await doWithDelay(timeout, () => {
-          currentPosition += delta;
-          setDelimerPercentPosition(currentPosition);
-        });
-      }
-      for (let i = 1; i <= PARTS; i += 1) {
-        await doWithDelay(timeout, () => {
-          currentPosition += delta;
-          setDelimerPercentPosition(currentPosition);
-        });
-      }
-      await doWithDelay(1000, () => {});
-      for (let i = 1; i <= PARTS; i += 1) {
-        await doWithDelay(timeout, () => {
-          currentPosition -= delta;
-          setDelimerPercentPosition(currentPosition);
-        });
-      }
-    }, 500);
-  };
-
-  return (
-    <BeforeAfterSlider
-      currentPercentPosition={delimerPersentPosition}
-      firstImage={before}
-      secondImage={after}
-      onVisible={animationDemo}
-      onChangePercentPosition={setDelimerPercentPosition}
-    />
-  );
-};
-
-const MDHeading = ({ children, id, tagType }) => {
-  const variantsStyle = {
-    h1: 'sm',
-    h2: 'sm',
-    h3: '18px',
-  };
-
-  return (
-    <Heading
-      as={tagType} // h1, h2, h3, h4, h5, h6
-      id={id}
-      size={variantsStyle[tagType] || 'sm'}
-      padding="20px 0 15px 0"
-      marginBottom="16px"
-    >
-      {children}
-    </Heading>
-  );
-};
-
-const MDCheckbox = ({ children }) => {
-  const stringg = 'checked';
-  return (
-    <Checkbox>
-      {`${children} - ${stringg}`}
-    </Checkbox>
-  );
-};
-
-const MDText = ({ children }) => {
-  const [haveHighlight, setHaveHighlight] = useState(false);
-  useEffect(() => {
-    if (children) {
-      // eslint-disable-next-line array-callback-return
-      children.map((child) => {
-        if (child && child.type && child.type.name === 'Code') {
-          setHaveHighlight(true);
-        }
-      });
-    } else {
-      console.log('something was wrong');
-    }
-  }, [children]);
-
-  return (
-    <Text size="l" className={haveHighlight ? 'text-highlight' : ''} overflow="auto !important" letterSpacing="0.05em" marginBottom="16px" fontWeight="400" lineHeight="24px">
-      {children}
-    </Text>
-  );
-};
-
-const MDLink = ({ children, href }) => (
-  <Link
-    href={href}
-    fontSize="15px"
-    color="blue.400"
-    fontWeight="700"
-    overflowWrap="anywhere"
-    target="_blank"
-    rel="noopener noreferrer"
-  >
-    {children}
-  </Link>
-);
-
-const MDTable = ({ children }) => (
-  <Box
-    as="div"
-    minW={{ base: '84vw', md: 'auto' }}
-    maxW={{ base: '84vw', md: 'auto' }}
-    // minW="100vw"
-    // maxW="100vw"
-    width="100%"
-    overflow="auto"
-  >
-    <Box
-      as="table"
-    >
-      {children}
-    </Box>
-
-  </Box>
-);
-
-const MDHr = () => (<Box d="inherit" />);
-
-const OnlyForBanner = ({ children, permission, cohortSession }) => {
-  const capabilities = permission.split(',');
-  console.log('md_permissions:', capabilities);
-  if (cohortSession.bc_id) {
-    // TODO: remove this param reassign
-    // eslint-disable-next-line no-param-reassign
-    cohortSession.user_capabilities = ['read_private_lesson', 'read_lesson', 'student'];
-  }
-
-  return (
-    <OnlyFor onlyMember withBanner cohortSession={cohortSession} capabilities={capabilities}>
-      {children}
-    </OnlyFor>
-  );
-};
+import CallToAction from '../CallToAction';
 
 const MarkDownParser = ({
   content, callToActionProps, withToc, frontMatter, titleRightSide,
@@ -226,13 +25,14 @@ const MarkDownParser = ({
 
   const newExerciseText = t('learnpack.new-exercise');
   const continueExerciseText = t('learnpack.continue-exercise');
+
   const {
     token, assetSlug, assetType, gitpod,
   } = callToActionProps;
 
-  const codeBlockBackticks = /\n``[^` ]*`/gm;
-  const simpleCodeblockBackticks = /```/gm;
-  const onlyforTag = /<onlyfor/gm;
+  const newLineBeforeCloseTag = /<\//gm;
+
+  const formatedContent = content.replace(newLineBeforeCloseTag, '\n$&');
 
   useEffect(() => {
     setLearnpackActions([
@@ -248,16 +48,6 @@ const MarkDownParser = ({
       },
     ]);
   }, [token, assetSlug, newExerciseText, continueExerciseText]);
-
-  // support for emoji shortcodes
-  // exapmle: :heart_eyes: -> 😍
-  const emojiSupport = (text) => text.replace(/:\w+:/gi, (name) => emoji.getUnicode(name));
-  const formatForCodeBlocks = content
-    .replace(codeBlockBackticks, '\n$&') // new line for codeBlocks and content
-    .replace(onlyforTag, '\n$&')
-    .replace(simpleCodeblockBackticks, '\n$&');
-
-  const contentFormated = emojiSupport(formatForCodeBlocks);
 
   return (
     <>
@@ -285,62 +75,32 @@ const MarkDownParser = ({
           <Toc content={content} />
         )}
       </ContentHeading>
-
-      {compiler(contentFormated, {
-        wrapper: null,
-        overrides: {
-          code: { component: Code },
-          p: { component: MDText },
-          a: { component: MDLink },
-          hr: { component: MDHr },
-          h1: {
-            component: MDHeading,
-            props: {
-              tagType: 'h2',
-            },
-          },
-          h2: {
-            component: MDHeading,
-            props: {
-              tagType: 'h2',
-            },
-          },
-          h3: {
-            component: MDHeading,
-            props: {
-              tagType: 'h3',
-            },
-          },
-          ul: {
-            props: { className: 'md-bullet' },
-          },
-          ol: {
-            props: { className: 'md-bullet' },
-          },
-          img: {
-            props: { className: 'MDImg' },
-          },
+      <ReactMarkdown
+      // gemoji plugin
+        remarkPlugins={[remarkGfm, remarkGemoji]}
+        rehypePlugins={[rehypeRaw]}
+        components={{
+          a: MDLink,
+          code: Code,
+          h1: ({ ...props }) => <MDHeading tagType="h2" {...props} />,
+          h2: ({ ...props }) => <MDHeading tagType="h2" {...props} />,
+          h3: ({ ...props }) => <MDHeading tagType="h3" {...props} />,
+          ul: ({ ...props }) => <ul className="md-bullet" {...props} />,
+          ol: ({ ...props }) => <ol className="md-bullet" {...props} />,
+          img: ({ ...props }) => <Img className="MDImg" {...props} />,
+          p: ({ ...props }) => <MDText {...props} />,
+          hr: ({ ...props }) => <MDHr {...props} />,
           BeforeAfter,
-          'before-after': {
-            component: BeforeAfter,
-          },
-          iframe: {
-            props: {
-              className: 'MDIframe',
-            },
-          },
-          table: {
-            component: MDTable,
-          },
-          onlyfor: {
-            component: OnlyForBanner,
-            props: {
-              cohortSession,
-            },
-          },
-        },
-        slugify: (str) => str.split(' ').join('-').toLowerCase(),
-      })}
+          'before-after': BeforeAfter,
+          iframe: ({ ...props }) => <iframe title={props.title || 'iframe-content'} className="MDIframe" {...props} />,
+          // table: {
+          //   component: MDTable,
+          // },
+          onlyfor: ({ ...props }) => <OnlyForBanner cohortSession={cohortSession} {...props} />,
+        }}
+      >
+        {formatedContent}
+      </ReactMarkdown>
     </>
   );
 };
@@ -358,62 +118,6 @@ MarkDownParser.defaultProps = {
   withToc: false,
   frontMatter: {},
   titleRightSide: null,
-};
-
-Code.propTypes = {
-  className: PropTypes.string,
-  children: PropTypes.node.isRequired,
-  inline: PropTypes.bool,
-};
-Code.defaultProps = {
-  className: '',
-  inline: false,
-};
-
-MDHeading.propTypes = {
-  children: PropTypes.node.isRequired,
-  id: PropTypes.string,
-  tagType: PropTypes.string,
-};
-
-MDHeading.defaultProps = {
-  id: '',
-  tagType: 'h2',
-};
-
-MDText.propTypes = {
-  children: PropTypes.node,
-};
-MDLink.propTypes = {
-  children: PropTypes.node.isRequired,
-  href: PropTypes.string.isRequired,
-};
-
-MDText.defaultProps = {
-  children: '',
-};
-
-BeforeAfter.propTypes = {
-  before: PropTypes.string.isRequired,
-  after: PropTypes.string.isRequired,
-};
-
-MDTable.propTypes = {
-  children: PropTypes.node.isRequired,
-};
-
-MDCheckbox.propTypes = {
-  children: PropTypes.node.isRequired,
-};
-
-OnlyForBanner.propTypes = {
-  children: PropTypes.node.isRequired,
-  permission: PropTypes.string,
-  cohortSession: PropTypes.objectOf(PropTypes.any),
-};
-OnlyForBanner.defaultProps = {
-  permission: '',
-  cohortSession: {},
 };
 
 export default MarkDownParser;
